Refactor ToolsMenu (/ToolsMenu.tsx, /ToolsMenu.module.css, and styles.tsx) to move the code related to packs into a dedicated module
