+++ utk_curio/frontend/urban-workflows/src/components/menus/nodes/toolsMenuPackPalette/registryBootstrap.ts
@@ -1,0 +1,7 @@
+import { getPaletteNodeTypes } from "../../../../registry";
+
+/** Changes whenever pack descriptors are registered — same signal as the outer palette rerender. */
+export function paletteDescriptorBootstrapKey(): string {
+    const types = getPaletteNodeTypes();
+    return `${types.length}|${types.map((d) => d.id).join(",")}`;
+}
