+++ utk_curio/frontend/urban-workflows/src/components/menus/nodes/toolsMenuPackPalette/uiConstants.ts
@@ -1,0 +1,7 @@
+/** Tooltip placement shared by built-in tools and pack-kind rows in the dock. */
+export type ToolsMenuTooltipSide = "left" | "right";
+
+export const OVERLAY_TRIGGER_DELAY_PROPS = {
+    show: 120,
+    hide: 10,
+};
