+++ utk_curio/frontend/urban-workflows/src/constants/nodeCategoryShortLabels.ts
@@ -1,0 +1,10 @@
+import type { NodeCategory } from "../registry/types";
+
+/** Short labels for node categories (palette chips, pack headers, etc.). */
+export const NODE_CATEGORY_SHORT_LABEL: Record<NodeCategory, string> = {
+    data: "Data",
+    computation: "Compute",
+    vis_grammar: "Viz",
+    vis_simple: "Chart",
+    flow: "Flow",
+};
